# SwaggerOpenAPI.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/users": {
            "post": {
                "operationId": "UsersController_create",
                "summary": "Criar novo usuário (apenas Admin e Superuser)",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "UsersController_findAll",
                "summary": "Listar todos os usuários",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "operationId": "UsersController_findOne",
                "summary": "Buscar usuário por ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "operationId": "UsersController_update",
                "summary": "Atualizar usuário",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "UsersController_remove",
                "summary": "Desativar usuário",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/permissions/user/{userId}": {
            "get": {
                "operationId": "PermissionsController_getUserPermissions",
                "summary": "Listar permissões de um usuário",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "permissions"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/access-logs/user/{userId}": {
            "get": {
                "operationId": "AccessLogController_findByUser",
                "summary": "Listar logs de acesso por usuário específico",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "access-logs"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Sistema de Controle de Acesso",
        "description": "API para gerenciamento de permissões e acesso a módulos",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateUserDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123",
                        "minLength": 6
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "SUPERUSER",
                            "ADMIN",
                            "USER"
                        ],
                        "example": "USER",
                        "description": "Papel do usuário no sistema"
                    }
                },
                "required": [
                    "email",
                    "name",
                    "password",
                    "role"
                ]
            },
            "UpdateUserDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "password": {
                        "type": "string",
                        "example": "newpassword123",
                        "minLength": 6
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "SUPERUSER",
                            "ADMIN",
                            "USER"
                        ],
                        "description": "Papel do usuário no sistema"
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    }
                }
            },
...
        }
    }
}